feat: melee double hit chance added

+++ Un1t (second iteration)/Assets/Scripts/EntitiesComponents/AttackData.cs
@@ -10,13 +10,18 @@ public readonly struct AttackData
     /// <remarks>
     /// This property equals 0 for every other entity
     /// </remarks>
-    public int XPDamage { get; }
+    public float XPDamage { get; }
     /// <summary>
     /// Transform of entity that can push player 
     /// </summary>
     public Transform AttackerTransform { get; }
     public float PushSpeed { get; }
-    public AttackData(float damage, DamageType damageType = DamageType.Physical, int xpDamage = 0, Transform transform = null, float pushSpeed = 0f)
+
+    public AttackData(float damage,
+                      DamageType damageType = DamageType.Physical,
+                      float xpDamage = 0,
+                      Transform transform = null,
+                      float pushSpeed = 0f)
     {
         Damage = damage;
         DamageType = damageType;
